refactor(http): extract build_response_headers_object to its own module (keep lib.rs under 2000-line cap)
